sin sqrt tanh atan add out, test=develop
