+++ python/paddle/__init__.py
@@ -120,20 +120,20 @@ import paddle.sysconfig
 # from .tensor.math import rsqrt   #DEFINE_ALIAS
 # from .tensor.math import scale   #DEFINE_ALIAS
 # from .tensor.math import sign   #DEFINE_ALIAS
-# from .tensor.math import sin   #DEFINE_ALIAS
-# from .tensor.math import sqrt   #DEFINE_ALIAS
+from .tensor.math import sin  #DEFINE_ALIAS
+from .tensor.math import sqrt  #DEFINE_ALIAS
 # from .tensor.math import square   #DEFINE_ALIAS
 # from .tensor.math import stanh   #DEFINE_ALIAS
 # from .tensor.math import sum   #DEFINE_ALIAS
 # from .tensor.math import sums   #DEFINE_ALIAS
-# from .tensor.math import tanh   #DEFINE_ALIAS
+from .tensor.math import tanh  #DEFINE_ALIAS
 # from .tensor.math import elementwise_sum   #DEFINE_ALIAS
 # from .tensor.math import max   #DEFINE_ALIAS
 # from .tensor.math import min   #DEFINE_ALIAS
 # from .tensor.math import mm   #DEFINE_ALIAS
 # from .tensor.math import div   #DEFINE_ALIAS
 # from .tensor.math import add   #DEFINE_ALIAS
-# from .tensor.math import atan   #DEFINE_ALIAS
+from .tensor.math import atan  #DEFINE_ALIAS
 # from .tensor.math import logsumexp   #DEFINE_ALIAS
 # from .tensor.math import inverse   #DEFINE_ALIAS
 # from .tensor.math import log1p   #DEFINE_ALIAS
