+++ python/paddle/tensor/__init__.py
@@ -69,7 +69,7 @@
 # from .math import abs   #DEFINE_ALIAS
 # from .math import acos   #DEFINE_ALIAS
 # from .math import asin   #DEFINE_ALIAS
-# from .math import atan   #DEFINE_ALIAS
+from .math import atan  #DEFINE_ALIAS
 # from .math import ceil   #DEFINE_ALIAS
 # from .math import cos   #DEFINE_ALIAS
 # from .math import cumsum   #DEFINE_ALIAS
@@ -98,13 +98,13 @@
 # from .math import rsqrt   #DEFINE_ALIAS
 # from .math import scale   #DEFINE_ALIAS
 # from .math import sign   #DEFINE_ALIAS
-# from .math import sin   #DEFINE_ALIAS
-# from .math import sqrt   #DEFINE_ALIAS
+from .math import sin  #DEFINE_ALIAS
+from .math import sqrt  #DEFINE_ALIAS
 # from .math import square   #DEFINE_ALIAS
 # from .math import stanh   #DEFINE_ALIAS
 # from .math import sum   #DEFINE_ALIAS
 # from .math import sums   #DEFINE_ALIAS
-# from .math import tanh   #DEFINE_ALIAS
+from .math import tanh  #DEFINE_ALIAS
 # from .math import elementwise_sum   #DEFINE_ALIAS
 # from .math import max   #DEFINE_ALIAS
 # from .math import min   #DEFINE_ALIAS
